Added support to Symfony Finder SplFileInfo
In Symfony 4, SplFileInfo provided by the Finder component requires its
constructor arguments to be strings
(see
https://github.com/symfony/symfony/commit/aaf2265203cfbefa006f0358077d8ade790ad203#diff-1c5661d50aeb4ac4fc0f6c0566fc2196L29).

This brings a few issues in phpspec (see
https://github.com/phpspec/phpspec/pull/1166#issuecomment-345616165),
this patch aims at fixing those issues.

Hopefully this patch isn't to farfetched, I assumed it was OK as there
are already a couple of patches for SplFileInfo.