Make two minor grammar updates to the README
While reading through the project README on github.com, I noticed a
small grammar mistake in it, and wanted to fix it. While making that
fix, I made a further slight change to the containing sentence to
improve the readability that much more. Nothing major. Just trying to
help where I can.