Make Prophecy Exceptions throwable
Constrains implementors of the interface to actually be Throwables. This prevents potential errors, and static analysis warnings when catching the interface (as the interface isn't itself throwable previously).

This wasn't possible in PHP 5, but we no longer support it. As a result getMessage does not need to be part of the interface, but also can't be mocked in tests